refactored to allow easy gathering of pagination summary in future
