+++ frontend/app/components/home/chapters.hbs
@@ -1,5 +1,3 @@
-<Home::Filter @chapters={{@chapters}} @tagFilters={{@selectedFilterTags}} as |results|>
-  {{#each results as |chapter|}}
-    <H5pListItem @chapter={{chapter}} />
-  {{/each}}
-</Home::Filter>
+{{#each @chapters as |chapter|}}
+  <H5pListItem @chapter={{chapter}} />
+{{/each}}
