--- frontend/app/components/home/filter.hbs
@@ -1,1 +1,0 @@
-{{yield this.results}}
