+++ frontend/app/controllers/home.js
@@ -3,7 +3,7 @@ import {inject as service} from '@ember/service';
 import {A} from '@ember/array';
 import {tracked} from '@glimmer/tracking';
 import {action} from '@ember/object';
-import {ChapterFilterTag} from '../pojos/chapter-filter-tag';
+import {ChapterFilterTag} from '../utils/chapter-filter-tag';
 
 export default class HomeController extends Controller {
   @service
@@ -11,6 +11,14 @@ export default class HomeController extends Controller {
 
   @tracked
   selectedFilterTags = A([]);
+
+  get chapters() {
+    if (this.selectedFilterTags.length > 0) {
+      return this.model
+        .filter((chapter) => this.selectedFilterTags.some((tag) => chapter.tags.includes(tag.name)));
+    }
+    return this.model;
+  }
 
 
   get tagsList() {
@@ -24,7 +32,6 @@ export default class HomeController extends Controller {
     });
     return allFilterTags.uniqBy('name');
   }
-
 
   @action
   clearAllTagFilters() {
