+++ frontend/app/templates/home.hbs
@@ -9,7 +9,7 @@
 
 <div class="container">
   <div class="row mt-5 ">
-    <Home::Chapters @chapters={{this.model}} @selectedFilterTags={{this.selectedFilterTags}}>
+    <Home::Chapters @chapters={{this.chapters}}>
     </Home::Chapters>
   </div>
