+++ frontend/app/utils/chapter-filter-tag.js
